feat(updated): add useAppMap hook
