+++ src/screens/AppMap/AppMap.tsx
@@ -1,140 +1,66 @@
-import React, {
-  useRef,
-  useState,
-  useCallback,
-  useLayoutEffect,
-  useMemo,
-  useEffect,
-} from 'react';
+import React, {useLayoutEffect, useEffect} from 'react';
 import {ClusterMap, ClusterMapShape, BottomMenu} from 'atoms';
-import {
-  createMarkerFromObject,
-  selectMapFilters,
-  selectTransformedData,
-  getMapMarkers,
-} from 'core/selectors';
-import {useSelector} from 'react-redux';
+import {createMarkerFromObject} from 'core/selectors';
 import {StyleProp, View} from 'react-native';
-import bbox from '@turf/bbox';
 
 import {styles, selectedPointStyle} from './styles';
-import {IMapFilter, IObject} from 'core/types';
-import MapBox, {
-  OnPressEvent,
-  SymbolLayerStyle,
-  RegionPayload,
-} from '@react-native-mapbox-gl/maps';
+import MapBox, {SymbolLayerStyle} from '@react-native-mapbox-gl/maps';
 import {
   AppMapBottomMenu,
   AppMapBottomSearchMenu,
   AppMapFilters,
   AppMapButtons,
 } from 'molecules';
-import {
-  useStatusBar,
-  useSearchList,
-  useFocusToUserLocation,
-  useBackHandler,
-  useColorScheme,
-  useTransformedData,
-  useAppMapAnalytics,
-  useBottomMenu,
-  useFindZoomForObjectInCluster,
-  useStaticCallback,
-} from 'core/hooks';
-import {useSafeAreaInsets} from 'react-native-safe-area-context';
-import {IProps} from './types';
+import {useBackHandler} from 'core/hooks';
 
-import {
-  Feature,
-  FeatureCollection,
-  Geometry,
-  Point,
-  Position,
-} from '@turf/helpers';
+import {FeatureCollection, Point} from '@turf/helpers';
+
 type SelecteMarker = ReturnType<typeof createMarkerFromObject>;
 
-import {mapService} from 'services/MapService';
 import {WINDOW_HEIGHT} from 'services/PlatformService';
-import Supercluster from 'supercluster';
-import {xorBy} from 'lodash';
-import {hapticFeedbackService} from 'services/HapticFeedbackService';
+import {useAppMap} from './hooks';
 
-export const AppMap = ({navigation}: IProps) => {
-  const sheme = useColorScheme();
-  const mapFilters = useSelector(selectMapFilters);
-  const appData = useSelector(selectTransformedData);
-
-  const camera = useRef<MapBox.Camera>(null);
-  const map = useRef<MapBox.MapView>(null);
-  const shapeSourceRef = useRef<MapBox.ShapeSource>(null);
-  const ignoreFitBounds = useRef(false);
-
-  const [selectedObject, setSelectedObject] = useState<null | IObject>(null);
-  const [selectedMarker, setSelectedMarker] = useState<SelecteMarker | null>(
-    () => createMarkerFromObject(null),
-  );
-
-  const [selectedFilters, setSelectedFilters] = useState<IMapFilter[]>([]);
-
-  const [markers, setMarkers] = useState(() =>
-    getMapMarkers(appData, selectedFilters),
-  );
-
-  const {getObject} = useTransformedData();
-  const {top} = useSafeAreaInsets();
-
-  const bounds = useMemo(() => {
-    if (markers) {
-      return mapService.getBoundsFromGeoJSON(markers, {
-        bottom: 70,
-        top: top + 20,
-      });
-    }
-
-    return null;
-  }, [markers, top]);
-
-  const {bottom} = useSafeAreaInsets();
+export const AppMap = () => {
   const {
-    data,
-    isHistoryVisible,
-    onTextChange,
-    addToHistory,
-    deleteAllFromHistory,
-    deleteFromHistory,
-    inputValue,
+    bounds,
+    ignoreFitBounds,
+    camera,
+    selectedObject,
+    openMenu,
+    isMenuOpened,
+    unselectObject,
+    isSearchMenuOpened,
+    closeSearchMenu,
     clearInput,
-  } = useSearchList({withLocation: true});
-
-  useStatusBar(sheme);
-
-  useAppMapAnalytics();
-
-  const {openMenu, closeMenu, isMenuOpened, ...menuProps} = useBottomMenu();
-  const {
-    openMenu: openSearchMenu,
-    closeMenu: closeSearchMenu,
-    isMenuOpened: isSearchMenuOpened,
-    ...searchMenuProps
-  } = useBottomMenu();
-
-  const onFilterSelect = useCallback(
-    (item: IMapFilter) => {
-      const newSelectedFilters = xorBy(selectedFilters, [item], 'categoryId');
-      const newMarkers = getMapMarkers(appData, newSelectedFilters);
-      setSelectedFilters(newSelectedFilters);
-      setMarkers(newMarkers);
-    },
-    [appData, selectedFilters],
-  );
-
-  const resetFilters = useCallback(() => {
-    const newSelectedFilters = [];
-    const newMarkers = getMapMarkers(appData, newSelectedFilters);
-    setSelectedFilters(newSelectedFilters);
-    setMarkers(newMarkers);
-  }, [appData]);
+    map,
+    unfocusUserLocation,
+    onShapePress,
+    onMapPress,
+    userLocationProps,
+    markers,
+    shapeSourceRef,
+    fitToClusterLeaves,
+    selectedMarker,
+    onMenuHideEnd,
+    menuProps,
+    bottom,
+    navigateToObjectDetails,
+    searchMenuProps,
+    onDeleteAllItems,
+    onDeleteItem,
+    inputValue,
+    isHistoryVisible,
+    data,
+    onSearchItemPress,
+    onTextChange,
+    isUserLocationFocused,
+    onShowLocationPress,
+    openSearchMenuAndPersistData,
+    onFilterSelect,
+    resetFilters,
+    selectedFilters,
+    mapFilters,
+  } = useAppMap();
 
   useLayoutEffect(() => {
     if (bounds) {
@@ -146,163 +72,11 @@ export const AppMap = ({navigation}: IProps) => {
     }
   }, [bounds]);
 
-  const unselectObject = useCallback(() => {
-    closeMenu();
-    setSelectedMarker(createMarkerFromObject(null));
-  }, [closeMenu]);
-
-  const onMapPress = useCallback(() => {
-    if (isMenuOpened()) {
-      unselectObject();
-    }
-  }, [isMenuOpened, unselectObject]);
-
   useEffect(() => {
     if (selectedObject) {
       openMenu();
     }
   }, [openMenu, selectedObject]);
-
-  const selectObjectAndOpenMenu = useCallback((object: IObject) => {
-    hapticFeedbackService.trigger();
-    setSelectedObject({...object});
-    setSelectedMarker(createMarkerFromObject(object));
-  }, []);
-
-  const onShapePress = useCallback(
-    async (objectId: string) => {
-      const itemData = getObject(objectId);
-
-      if (itemData) {
-        const currentZoom = await map.current?.getZoom();
-
-        const coordinates = [itemData.location!.lon!, itemData.location!.lat!];
-        camera.current?.setCamera({
-          centerCoordinate: coordinates,
-          zoomLevel: currentZoom,
-          animationDuration: 500,
-        });
-        selectObjectAndOpenMenu(itemData);
-      }
-    },
-    [getObject, selectObjectAndOpenMenu],
-  );
-
-  const navigateToObjectDetails = useCallback(
-    ({id}: IObject) => {
-      unselectObject();
-      navigation.push('ObjectDetails', {objectId: id});
-    },
-    [unselectObject, navigation],
-  );
-
-  const {findZoomForObjectInCluster} = useFindZoomForObjectInCluster();
-
-  const moveCameraToSearchedObject = useCallback(
-    async (
-      object: IObject,
-      cluster: Supercluster<Supercluster.AnyProps, Supercluster.AnyProps>,
-      clusterBounds,
-    ) => {
-      const location = object.location;
-      const coordinates = location ? [location.lon!, location.lat!] : null;
-
-      if (coordinates) {
-        const visibleClusters = cluster.getClusters(clusterBounds, 4);
-
-        const currentZoom = await map.current?.getZoom();
-
-        if (visibleClusters && currentZoom) {
-          const zoomLevel = findZoomForObjectInCluster(
-            cluster,
-            object,
-            visibleClusters,
-            currentZoom,
-          );
-
-          if (zoomLevel) {
-            camera.current?.setCamera({
-              centerCoordinate: coordinates,
-              zoomLevel: zoomLevel,
-              animationDuration: 600,
-            });
-          }
-        }
-      }
-    },
-    [findZoomForObjectInCluster],
-  );
-
-  const onSearchItemPress = useCallback(
-    (object: IObject) => {
-      let newFitlters = selectedFilters;
-      let newMarkers = markers;
-
-      if (selectedFilters.length) {
-        ignoreFitBounds.current = true;
-
-        newFitlters = [];
-        newMarkers = getMapMarkers(appData, newFitlters);
-
-        setSelectedFilters(newFitlters);
-        setMarkers(newMarkers);
-      }
-
-      closeSearchMenu();
-      const clusterBounds = bbox(newMarkers);
-      const cluster = new Supercluster({
-        radius: 40,
-        maxZoom: 14,
-      }).load(newMarkers?.features!);
-
-      moveCameraToSearchedObject(object, cluster, clusterBounds);
-
-      addToHistory(object);
-      selectObjectAndOpenMenu(object);
-      clearInput();
-    },
-    [
-      addToHistory,
-      appData,
-      clearInput,
-      closeSearchMenu,
-      markers,
-      moveCameraToSearchedObject,
-      selectObjectAndOpenMenu,
-      selectedFilters,
-    ],
-  );
-
-  const onDeleteItem = useCallback(
-    (searchItem: IObject) => {
-      deleteFromHistory(searchItem);
-    },
-    [deleteFromHistory],
-  );
-
-  const onDeleteAllItems = useCallback(() => {
-    deleteAllFromHistory();
-  }, [deleteAllFromHistory]);
-
-  const onMenuHideEnd = useStaticCallback(() => {
-    setSelectedObject(null);
-
-    if (selectedMarker) {
-      setSelectedMarker(createMarkerFromObject(null));
-    }
-  }, [selectedMarker]);
-
-  const openSearchMenuAndPersistData = useCallback(() => {
-    closeMenu();
-    openSearchMenu();
-  }, [closeMenu, openSearchMenu]);
-
-  const {
-    focusToUserLocation,
-    setIsUserLocationFocused,
-    isUserLocationFocused,
-    ...userLocationProps
-  } = useFocusToUserLocation(camera);
 
   useBackHandler(() => {
     if (isMenuOpened()) {
@@ -318,45 +92,6 @@ export const AppMap = ({navigation}: IProps) => {
 
     return false;
   });
-
-  const fitToClusterLeaves = useCallback(async (event: OnPressEvent) => {
-    const {features} = event;
-    const isCluster = features[0]?.properties?.cluster;
-
-    if (isCluster) {
-      const cluster = features[0] as Feature<Geometry, {cluster_id: number}>;
-      const {
-        geometry: {coordinates},
-      } = cluster;
-
-      const zoom = await shapeSourceRef.current?.getClusterExpansionZoom(
-        cluster,
-      );
-      camera.current?.setCamera({
-        centerCoordinate: coordinates as Position,
-        zoomLevel: zoom,
-        animationDuration: 200,
-      });
-    }
-  }, []);
-
-  const onShowLocationPress = useCallback(async () => {
-    const currentZoom = (await map.current?.getZoom()) || null;
-
-    focusToUserLocation(bounds, currentZoom);
-  }, [bounds, focusToUserLocation]);
-
-  const unfocusUserLocation = useCallback(
-    (feature: Feature<Point, RegionPayload>) => {
-      const {
-        properties: {isUserInteraction},
-      } = feature;
-      if (isUserInteraction) {
-        setIsUserLocationFocused(false);
-      }
-    },
-    [setIsUserLocationFocused],
-  );
 
   return (
     <View style={styles.container}>
