+++ src/screens/AppMap/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useAppMap} from './useAppMap';
